backend/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
idna==3.3
Pillow==8.3.2
psycopg2-binary==2.8.6
PyJWT==2.1.0
python-decouple==3.5
pytz==2021.1
redis==3.5.3
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.7